move Details related classes to own package
