+++ app/src/main/java/com/derek_s/hubble_gallery/internal/di/ActivityComponent.java
@@ -1,9 +1,9 @@
 package com.derek_s.hubble_gallery.internal.di;
 
-import com.derek_s.hubble_gallery.act_main.ActMain;
 import com.derek_s.hubble_gallery.base.ActBase;
 import com.derek_s.hubble_gallery.base.FragBase;
-import com.derek_s.hubble_gallery.ui.activities.ActDetails;
+import com.derek_s.hubble_gallery.detailspage.DetailsActivity;
+import com.derek_s.hubble_gallery.home.ActHome;
 import com.derek_s.hubble_gallery.ui.activities.ActImageViewer;
 import com.derek_s.hubble_gallery.ui.activities.ActWelcome;
 
@@ -19,9 +19,9 @@ public interface ActivityComponent {
 
   void inject(ActBase app);
 
-  void inject(ActDetails app);
+  void inject(DetailsActivity app);
 
-  void inject(ActMain app);
+  void inject(ActHome app);
 
   void inject(ActImageViewer app);
